requirements.txt
fastapi
google-genai
jinja2
pytest
python-dotenv
python-multipart
pyyaml
uvicorn